requirements.txt
nemo_toolkit[asr]
fastapi
python-multipart
librosa
soundfile
uvicorn[standard]
scipy